Гончарі: таблиця без стелі long — Владік більше не «перший», коли попереду Микола
Коло писало в таблицю min(глеки, 9,2·10¹⁸). Микола (3,6·10²¹) і Владік (4,2·10¹⁹)
обоє вперлись у стелю, зрівнялись, і рядок суперника кожному казав «ти перший».
Соло-рахунок тепер double від гри до бази (SQLite сам кладе REAL у score),
таблиця й рекорд читають його як double.

+++ src/Hlechyky/Games/Contracts.cs
@@ -141,7 +141,7 @@ public interface IRoomContext
     /// Соло-результат у таблицю (порядок — Info.Score). <paramref name="attempts"/> — для щоденних, що міряють
     /// і час, і спроби: тоді value — мілісекунди, а спроби йдуть окремо, а не вгадуються з величини числа.
     /// </summary>
-    void Score(int seat, long value, int? attempts = null);
+    void Score(int seat, double value, int? attempts = null);
     /// <summary>Черепки поза стандартною виплатою за партію (переможець «Скільки?», конкурс). Проходить через стелі економіки.</summary>
     void Award(int seat, int shards, string reason);
 }
@@ -274,7 +274,7 @@ public sealed record RoomFinishedEvent(
     DateTimeOffset StartedAt, DateTimeOffset FinishedAt,
     int Moves);                         // скільки Act прийнято за партію (0 для реалтайму)
 
-public sealed record SoloScoreEvent(string GameId, string Nick, long Score, ScoreOrder Order, string? Key, DateTimeOffset At,
+public sealed record SoloScoreEvent(string GameId, string Nick, double Score, ScoreOrder Order, string? Key, DateTimeOffset At,
     int? Attempts = null);
 
 public sealed record AwardEvent(string GameId, string RoomId, string Nick, int Shards, string Reason);
